Break down the old 'artworks.js' into smaller components.
