+++ src/exceptions/artworks-view-error.js
@@ -1,0 +1,10 @@
+class ArtworksViewError extends Error {
+  constructor(position, data, ...options) {
+    super(...options);
+    this.name = 'ArtworksError';
+    this.position = position;
+    this.data = data;
+  }
+}
+
+module.exports = ArtworksViewError;
