+++ src/exceptions/unsupported-record-error.js
@@ -1,0 +1,3 @@
+class UnsupportedRecordError extends Error {}
+
+module.exports = UnsupportedRecordError;
